+++ apps/cli/cli_common.c
@@ -520,7 +520,7 @@ cli_quit(clicon_handle h,
 	  cvec         *vars, 
 	  cvec         *argv)
 {
-    cli_set_exiting(h, 1);
+    cligen_exiting_set(cli_cligen(h), 1);
     return 0;
 }
 int cli_quitv(clicon_handle h, cvec *vars, cvec *argv)